+++ src/Shipment/resources/database/migrations/2023_03_28_195634_add_extend_the_shipments_table.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 return new class () extends Migration {
     public function up()
     {
         Schema::table('shipments', function (Blueprint $table) {
             $table->decimal('carrier_cost', 15, 4)->nullable()->after('configuration');
+            $table->string('label_url', 1024)->nullable()->after('configuration');
+            $table->longText('label_base64')->nullable()->after('configuration');
         });
     }
 
@@ -17,6 +20,12 @@ return new class () extends Migration {
     {
         Schema::table('shipments', function (Blueprint $table) {
             $table->dropColumn('carrier_cost');
+        });
+        Schema::table('shipments', function (Blueprint $table) {
+            $table->dropColumn('label_url');
+        });
+        Schema::table('shipments', function (Blueprint $table) {
+            $table->dropColumn('label_base64');
         });
     }
 };